web/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
python-dotenv==1.0.0
werkzeug==3.0.1
docker==7.0.0
